Fix volume reference issues in create_os_instance
During testing, create_os_instance was querying for non-existant volume
names. That query seems to be redundant given that we have the UUID from
the migate plugin call. That is changed to remove the lookup.

When investigating, we also noticed that the metadata set
commands seems to use a non-existant variable. That is changed to use
the first UUID from the UUID list.